+++ crates/sindri-scene/src/extract/mesh.rs
@@ -73,6 +73,12 @@ impl SceneExtractor {
                         .world
                         .ok_or(SceneExtractError::MissingWorldCamera)?
                         .view_projection,
+                    position: cameras
+                        .world
+                        .ok_or(SceneExtractError::MissingWorldCamera)?
+                        .view
+                        .inverse()
+                        .transform_point3(glam::Vec3::ZERO),
                 },
                 command,
             ));